// proto/interfaces/present.ts
export interface OrderDto {
  orderId: string;
  medicines: MedicineDto[];
  geolocation: GeoJSONPoint | undefined;
  client: ClientDto | undefined;
}

export interface ClientDto {
  name: string;
  address: string;
  phone: string;
  identity: string;
}

export interface MedicineDto {
  medicineId: string;
  quantity: number;
  name: string;
  barcode: string;
  price: number;
}

export interface GeoJSONPoint {
  latitude: number;
  longitude: number;
}

function createBaseOrderDto(): OrderDto {
  return { orderId: "", medicines: [], geolocation: undefined, client: undefined };
}

function createBaseClientDto(): ClientDto {
  return { name: "", address: "", phone: "", identity: "" };
}

export const OrderDto: MessageFns<OrderDto> = {
  encode(message: OrderDto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.orderId !== "") {
      writer.uint32(10).string(message.orderId);
    }
    for (const v of message.medicines) {
      MedicineDto.encode(v!, writer.uint32(18).fork()).join();
    }
    if (message.geolocation !== undefined) {
      GeoJSONPoint.encode(message.geolocation, writer.uint32(26).fork()).join();
    }
    if (message.client !== undefined) {
      ClientDto.encode(message.client, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): OrderDto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseOrderDto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.orderId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.medicines.push(MedicineDto.decode(reader, reader.uint32()));
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.geolocation = GeoJSONPoint.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.client = ClientDto.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): OrderDto {
    return {
      orderId: isSet(object.orderId)
        ? globalThis.String(object.orderId)
        : isSet(object.order_id)
        ? globalThis.String(object.order_id)
        : "",
      medicines: globalThis.Array.isArray(object?.medicines)
        ? object.medicines.map((e: any) => MedicineDto.fromJSON(e))
        : [],
      geolocation: isSet(object.geolocation) ? GeoJSONPoint.fromJSON(object.geolocation) : undefined,
      client: isSet(object.client) ? ClientDto.fromJSON(object.client) : undefined,
    };
  },

  toJSON(message: OrderDto): unknown {
    const obj: any = {};
    if (message.orderId !== "") {
      obj.orderId = message.orderId;
    }
    if (message.medicines?.length) {
      obj.medicines = message.medicines.map((e) => MedicineDto.toJSON(e));
    }
    if (message.geolocation !== undefined) {
      obj.geolocation = GeoJSONPoint.toJSON(message.geolocation);
    }
    if (message.client !== undefined) {
      obj.client = ClientDto.toJSON(message.client);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<OrderDto>, I>>(base?: I): OrderDto {
    return OrderDto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<OrderDto>, I>>(object: I): OrderDto {
    const message = createBaseOrderDto();
    message.orderId = object.orderId ?? "";
    message.medicines = object.medicines?.map((e) => MedicineDto.fromPartial(e)) || [];
    message.geolocation = (object.geolocation !== undefined && object.geolocation !== null)
      ? GeoJSONPoint.fromPartial(object.geolocation)
      : undefined;
    message.client = (object.client !== undefined && object.client !== null)
      ? ClientDto.fromPartial(object.client)
      : undefined;
    return message;
  },
};

export const ClientDto: MessageFns<ClientDto> = {
  encode(message: ClientDto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.name !== "") writer.uint32(10).string(message.name);
    if (message.address !== "") writer.uint32(18).string(message.address);
    if (message.phone !== "") writer.uint32(26).string(message.phone);
    if (message.identity !== "") writer.uint32(34).string(message.identity);
    return writer;
  },
  decode(input: BinaryReader | Uint8Array, length?: number): ClientDto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseClientDto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: message.name = reader.string(); break;
        case 2: message.address = reader.string(); break;
        case 3: message.phone = reader.string(); break;
        case 4: message.identity = reader.string(); break;
        default: reader.skip(tag & 7); break;
      }
    }
    return message;
  },
  fromJSON(object: any): ClientDto {
    return {
      name: isSet(object.name) ? String(object.name) : "",
      address: isSet(object.address) ? String(object.address) : "",
      phone: isSet(object.phone) ? String(object.phone) : "",
      identity: isSet(object.identity) ? String(object.identity) : "",
    };
  },
  toJSON(message: ClientDto): unknown {
    return { ...message };
  },
  create<I extends Exact<DeepPartial<ClientDto>, I>>(base?: I): ClientDto {
    return ClientDto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ClientDto>, I>>(object: I): ClientDto {
    const message = createBaseClientDto();
    message.name = object.name ?? "";
    message.address = object.address ?? "";
    message.phone = object.phone ?? "";
    message.identity = object.identity ?? "";
    return message;
  },
};

function createBaseMedicineDto(): MedicineDto {
  return { medicineId: "", quantity: 0, name: "", barcode: "", price: 0 };
}

export const MedicineDto: MessageFns<MedicineDto> = {
  encode(message: MedicineDto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.medicineId !== "") writer.uint32(10).string(message.medicineId);
    if (message.quantity !== 0) writer.uint32(16).uint32(message.quantity);
    if (message.name !== "") writer.uint32(26).string(message.name);
    if (message.barcode !== "") writer.uint32(34).string(message.barcode);
    if (message.price !== 0) writer.uint32(41).double(message.price);
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MedicineDto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMedicineDto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: message.medicineId = reader.string(); break;
        case 2: message.quantity = reader.uint32(); break;
        case 3: message.name = reader.string(); break;
        case 4: message.barcode = reader.string(); break;
        case 5: message.price = reader.double(); break;
        default: reader.skip(tag & 7); break;
      }
    }
    return message;
  },

  fromJSON(object: any): MedicineDto {
    return {
      medicineId: isSet(object.medicineId)
        ? globalThis.String(object.medicineId)
        : isSet(object.medicine_id)
        ? globalThis.String(object.medicine_id)
        : "",
      quantity: isSet(object.quantity) ? globalThis.Number(object.quantity) : 0,
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      barcode: isSet(object.barcode) ? globalThis.String(object.barcode) : "",
      price: isSet(object.price) ? globalThis.Number(object.price) : 0,
    };
  },

  toJSON(message: MedicineDto): unknown {
    const obj: any = {};
    if (message.medicineId !== "") {
      obj.medicineId = message.medicineId;
    }
    if (message.quantity !== 0) {
      obj.quantity = Math.round(message.quantity);
    }
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.barcode !== "") {
      obj.barcode = message.barcode;
    }
    if (message.price !== 0) {
      obj.price = message.price;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MedicineDto>, I>>(base?: I): MedicineDto {
    return MedicineDto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MedicineDto>, I>>(object: I): MedicineDto {
    const message = createBaseMedicineDto();
    message.medicineId = object.medicineId ?? "";
    message.quantity = object.quantity ?? 0;
    message.name = object.name ?? "";
    message.barcode = object.barcode ?? "";
    message.price = object.price ?? 0;
    return message;
  },
};

function createBaseGeoJSONPoint(): GeoJSONPoint {
  return { latitude: 0, longitude: 0 };
}

export const GeoJSONPoint: MessageFns<GeoJSONPoint> = {
  encode(message: GeoJSONPoint, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.latitude !== 0) {
      writer.uint32(9).double(message.latitude);
    }
    if (message.longitude !== 0) {
      writer.uint32(17).double(message.longitude);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GeoJSONPoint {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGeoJSONPoint();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 9) {
            break;
          }

          message.latitude = reader.double();
          continue;
        }
        case 2: {
          if (tag !== 17) {
            break;
          }

          message.longitude = reader.double();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GeoJSONPoint {
    return {
      latitude: isSet(object.latitude) ? globalThis.Number(object.latitude) : 0,
      longitude: isSet(object.longitude) ? globalThis.Number(object.longitude) : 0,
    };
  },

  toJSON(message: GeoJSONPoint): unknown {
    const obj: any = {};
    if (message.latitude !== 0) {
      obj.latitude = message.latitude;
    }
    if (message.longitude !== 0) {
      obj.longitude = message.longitude;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GeoJSONPoint>, I>>(base?: I): GeoJSONPoint {
    return GeoJSONPoint.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GeoJSONPoint>, I>>(object: I): GeoJSONPoint {
    const message = createBaseGeoJSONPoint();
    message.latitude = object.latitude ?? 0;
    message.longitude = object.longitude ?? 0;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}